# docs/swagger.yaml
definitions:
  common_respone.HTTPError:
    properties:
      error:
        example: any error
        type: string
    type: object
  common_respone.HTTPSuccess:
    properties:
      success:
        type: boolean
    type: object
  friendship.RequestFriend:
    properties:
      friends:
        items:
          type: string
        type: array
    required:
    - friends
    type: object
  friendship.RequestListFriends:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  friendship.RequestReceiveUpdate:
    properties:
      sender:
        type: string
      text:
        type: string
    required:
    - sender
    - text
    type: object
  friendship.RequestUpdate:
    properties:
      requestor:
        type: string
      target:
        type: string
    required:
    - requestor
    - target
    type: object
  friendship.ResponeListFriends:
    properties:
      count:
        type: integer
      friends:
        items:
          type: string
        type: array
      success:
        type: boolean
    type: object
  friendship.ResponeReceiveUpdate:
    properties:
      recipients:
        items:
          type: string
        type: array
      success:
        type: boolean
    type: object
  user.RequestCreateUser:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  user.ResponeListUser:
    properties:
      count:
        type: integer
      list_users:
        items:
          type: string
        type: array
    required:
    - count
    - list_users
    type: object
info:
  contact: {}
paths:
  /add-friends:
    post:
      description: Create a friend connection between two email addresses.
      parameters:
      - description: RequestCreateUser
        in: body
        name: friends
        required: true
        schema:
          $ref: '#/definitions/friendship.RequestFriend'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common_respone.HTTPSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common_respone.HTTPError'
      summary: Make Friend Connection
      tags:
      - Friendship
  /block:
    post:
      description: Block updates from an email address.
      parameters:
      - description: Requestor and Target to block update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/friendship.RequestUpdate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common_respone.HTTPSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common_respone.HTTPError'
      summary: Block Subscribe to update an user
      tags:
      - Friendship
  /create-user:
    post:
      description: Create A New User
      parameters:
      - description: RequestCreateUser
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/user.RequestCreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common_respone.HTTPSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common_respone.HTTPError'
      summary: Create A New User
      tags:
      - User
  /get-list-friends:
    post:
      description: Retrieve the friends list for an email address.
      parameters:
      - description: RequestListFriends
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/friendship.RequestListFriends'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/friendship.ResponeListFriends'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common_respone.HTTPError'
      summary: Get Friends List
      tags:
      - Friendship
  /get-list-users-receive-update:
    post:
      description: Retrieve all email addresses that can receive updates from an email address.
      parameters:
      - description: Sender and Text
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/friendship.RequestReceiveUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/friendship.ResponeReceiveUpdate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common_respone.HTTPError'
      summary: Get Users Receive Update
      tags:
      - Friendship
  /get-mutual-list-friends:
    post:
      description: Retrieve the common friends list between two email addresses.
      parameters:
      - description: RequestFriend
        in: body
        name: friends
        required: true
        schema:
          $ref: '#/definitions/friendship.RequestFriend'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/friendship.ResponeListFriends'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common_respone.HTTPError'
      summary: Get Mutual Friends List
      tags:
      - Friendship
  /list-users:
    get:
      description: Get list users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.ResponeListUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common_respone.HTTPError'
      summary: List users
      tags:
      - User
  /subscribe:
    post:
      description: Subscribe to updates from an email address.
      parameters:
      - description: Requestor and Target to subscribe update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/friendship.RequestUpdate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common_respone.HTTPSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common_respone.HTTPError'
      summary: Subscribe update an user
      tags:
      - Friendship
swagger: "2.0"